Moving the configuration cruft out of the LinterPlugin code
