+++ src/main/scala/LinterConfig.scala
@@ -1,0 +1,18 @@
+package com.foursquare.lint
+
+import com.typesafe.config.ConfigFactory 
+
+/**
+ * <p>Configuration class for running (or not) different lint checks</p>
+ * <p>As a library, linter defaults to including the src/main/resources/reference.conf settings
+ * which are override-able by an including application.</p>
+ */
+object LinterConfig {
+  
+  val config = ConfigFactory.load("linter")
+  config.checkValid(ConfigFactory.defaultReference(), "linter")
+  
+  val packageWildcardWhitelistCheckEnabled = config.getBoolean("linter.package.wildcard.whitelist.checkEnabled")
+  val packageWildcardWhitelistPackages = config.getStringList("linter.package.wildcard.whitelist.packages")
+  val packageWildcardWhitelistSeverity = LinterSeverity.withName(config.getString("linter.package.wildcard.whitelist.severity"))
+}
